+++ adaptor/transformer/testdata/goja-transformer.js
@@ -1,0 +1,1 @@
+function f(doc) { return doc }